Time: 88 ms (18.73%), Space: 54.8 MB (98.42%) - LeetHub

+++ 2665-minimum-time-to-repair-cars/2665-minimum-time-to-repair-cars.java
@@ -6,7 +6,7 @@ class Solution {
             long mid = (minimumTime + maximumTime) / 2;
             long sum = 0;
             for(int index = 0; index < ranks.length; index++){
-                sum += (long) Math.floor(Math.sqrt((double) mid / ranks[index]));
+                sum += (long) Math.floor(Math.sqrt(mid / ranks[index]));
             }
             if (sum >= cars) {
                 maximumTime = mid;
